+++ packages/react/src/components/avatar/styles.ts
@@ -1,12 +1,12 @@
 import * as Avatar from '@radix-ui/react-avatar'
 
-import { styled } from '../../../styles'
+import { styled } from '../../styles'
 
 export const AvatarContainer = styled(Avatar.Root, {
 	borderRadius: '$full',
 	display: 'inline-block',
-	width: '$12',
-	height: '$12',
+	width: '$16',
+	height: '$16',
 	overflow: 'hidden',
 })