+++ packages/react/src/components/box.tsx
@@ -1,6 +1,6 @@
 import { ComponentProps, ElementType } from 'react'
 
-import { styled } from '../../styles'
+import { styled } from '../styles'
 
 export const Box = styled('div', {
 	padding: '$6',