+++ packages/react/src/components/heading.tsx
@@ -1,6 +1,6 @@
 import { ComponentProps, ElementType } from 'react'
 
-import { styled } from '../../styles'
+import { styled } from '../styles'
 
 export const Heading = styled('h2', {
 	fontFamily: '$default',