+++ packages/react/src/components/text-input/index.tsx
@@ -1,18 +1,21 @@
-import { ComponentProps } from 'react'
+import { ComponentProps, ElementRef, forwardRef } from 'react'
 
 import { Input, Prefix, TextInputContainer } from './styles'
 
 export interface TextInputProps extends ComponentProps<typeof Input> {
 	prefix?: string
+	containerProps?: ComponentProps<typeof TextInputContainer>
 }
 
-export function TextInput({ prefix, ...props }: TextInputProps) {
-	return (
-		<TextInputContainer>
-			{!!prefix && <Prefix>{prefix}</Prefix>}
-			<Input {...props} />
-		</TextInputContainer>
-	)
-}
+export const TextInput = forwardRef<ElementRef<typeof Input>, TextInputProps>(
+	({ prefix, containerProps, ...props }: TextInputProps, ref) => {
+		return (
+			<TextInputContainer {...containerProps}>
+				{!!prefix && <Prefix>{prefix}</Prefix>}
+				<Input ref={ref} {...props} />
+			</TextInputContainer>
+		)
+	},
+)
 
 TextInput.displayName = 'TextInput'