+++ packages/react/src/index.tsx
@@ -7,3 +7,5 @@ export * from './components/text-input'
 export * from './components/text-area'
 export * from './components/checkbox'
 export * from './components/multi-step'
+
+export * from './styles'