+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rcrdk-ignite-ui/react
 
+## 2.0.1
+
+### Patch Changes
+
+- - Fixed components styles.
+  - Add ref to TextInput component.
+  - Fixed styles folder location.
+
 ## 2.0.0
 
 ### Major Changes